fix: resolve the issue of generated titles missing id
