+++ console/packages/editor/src/utils/anchor.ts
