+++ console/packages/editor/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./delete-node";
+export * from "./anchor";
